requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangGraph and AI
langgraph==0.0.60
langchain==0.1.0
langchain-anthropic==0.1.1
anthropic==0.8.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# Rate Limiting
slowapi==0.1.9

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Circuit Breaker
pybreaker==1.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.12.0
ruff==0.1.7
mypy==1.7.1
